agg/requirements.txt
fastapi
uvicorn
requests
asyncio
httpx
python-dotenv
